# walking-safely/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Walking Safely API",
        "description": "API para o aplicativo Walking Safely - Navegação segura com análise de risco em tempo real",
        "contact": {
            "name": "Walking Safely Team",
            "email": "suporte@walkingsafely.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de Desenvolvimento"
        },
        {
            "url": "https://api.walkingsafely.com/api",
            "description": "Servidor de Produção"
        }
    ],
    "paths": {
...
        "/occurrences": {
            "get": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Listar ocorrências",
                "description": "Retorna lista paginada de ocorrências com filtros opcionais",
                "operationId": "listOccurrences",
                "parameters": [
                    {
                        "name": "region_id",
                        "in": "query",
                        "description": "Filtrar por região",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "crime_type_id",
                        "in": "query",
                        "description": "Filtrar por tipo de crime",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "description": "Filtrar por severidade",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "low",
                                "medium",
                                "high",
                                "critical"
                            ]
                        }
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "description": "Data inicial (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "description": "Data final (YYYY-MM-DD)",
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "days",
                        "in": "query",
                        "description": "Últimos N dias",
                        "schema": {
                            "type": "integer",
                            "maximum": 365,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Latitude para busca por proximidade",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Longitude para busca por proximidade",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Raio de busca em metros (padrão: 1000)",
                        "schema": {
                            "type": "number",
                            "maximum": 50000,
                            "minimum": 100
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Itens por página (padrão: 15)",
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de ocorrências",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Occurrence"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/components/schemas/PaginationMeta"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Criar ocorrência",
                "description": "Registra uma nova ocorrência criminal. Coordenadas devem estar a até 100m da localização do usuário. Limite de 5 reportes por hora por usuário.",
                "operationId": "createOccurrence",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "latitude",
                                    "longitude",
                                    "crime_type_id",
                                    "severity",
                                    "user_location"
                                ],
                                "properties": {
                                    "latitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -23.5505
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "float",
                                        "example": -46.6333
                                    },
                                    "crime_type_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "severity": {
                                        "type": "string",
                                        "enum": [
                                            "low",
                                            "medium",
                                            "high",
                                            "critical"
                                        ],
                                        "example": "medium"
                                    },
                                    "timestamp": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2025-01-05T10:30:00Z"
                                    },
                                    "user_location": {
                                        "required": [
                                            "latitude",
                                            "longitude"
                                        ],
                                        "properties": {
                                            "latitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -23.5506
                                            },
                                            "longitude": {
                                                "type": "number",
                                                "format": "float",
                                                "example": -46.6334
                                            }
                                        },
                                        "type": "object"
                                    },
                                    "metadata": {
                                        "properties": {
                                            "description": {
                                                "type": "string",
                                                "maxLength": 500,
                                                "example": "Assalto a mão armada"
                                            }
                                        },
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Ocorrência criada com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Occurrence"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Ocorrência registrada com sucesso"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação ou localização inválida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    },
                    "429": {
                        "description": "Limite de reportes excedido",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "rate_limit_exceeded"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Limite de 5 reportes por hora excedido"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/occurrences/{id}": {
            "get": {
                "tags": [
                    "Occurrences"
                ],
                "summary": "Obter ocorrência",
                "description": "Retorna os detalhes de uma ocorrência específica",
                "operationId": "getOccurrence",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da ocorrência",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da ocorrência",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Occurrence"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Ocorrência não encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Occurrence": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "latitude": {
                        "type": "number",
                        "format": "float",
                        "example": -23.5505
                    },
                    "longitude": {
                        "type": "number",
                        "format": "float",
                        "example": -46.6333
                    },
                    "crime_type_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "crime_type": {
                        "type": "string",
                        "example": "Assalto"
                    },
                    "severity": {
                        "type": "string",
                        "enum": [
                            "low",
                            "medium",
                            "high",
                            "critical"
                        ],
                        "example": "medium"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "expired",
                            "rejected",
                            "merged"
                        ],
                        "example": "active"
                    },
                    "source": {
                        "type": "string",
                        "enum": [
                            "collaborative",
                            "official"
                        ],
                        "example": "collaborative"
                    },
                    "timestamp": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "confidence_score": {
                        "type": "integer",
                        "example": 3
                    },
                    "region_id": {
                        "type": "integer",
                        "example": 5,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "Error": {
                "properties": {
                    "error": {
                        "type": "string",
                        "example": "validation_error"
                    },
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos"
                    }
                },
                "type": "object"
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Os dados fornecidos são inválidos"
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "type": "object"
            },
            "PaginationMeta": {
                "properties": {
                    "current_page": {
                        "type": "integer",
                        "example": 1
                    },
                    "from": {
                        "type": "integer",
                        "example": 1
                    },
                    "last_page": {
                        "type": "integer",
                        "example": 10
                    },
                    "per_page": {
                        "type": "integer",
                        "example": 15
                    },
                    "to": {
                        "type": "integer",
                        "example": 15
                    },
                    "total": {
                        "type": "integer",
                        "example": 150
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}